pyproject.toml
[project]
name = "oauth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.133.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.13.1",
    "uvicorn>=0.41.0",
]
